move dto folder to correct module
